Backend - updated quantity to allow decimals

+++ RecipeBackend/DTOs/IngredientDtos.cs
@@ -1,7 +1,9 @@
+using System.Security.Cryptography.X509Certificates;
+
 public class CreateInventoryIngredientDto
 {
     public int UserId { get; set; }
-    public int? Quantity { get; set; }
+    public decimal? Quantity { get; set; }
     public int? QuantityUnitId { get; set; }
     public int IngredientId { get; set; }
 }
@@ -10,7 +12,7 @@ public class CreateListIngredientDto
 {
     public int UserId { get; set; }
     public int IngredientId { get; set; }
-    public int? Quantity { get; set; }
+    public decimal? Quantity { get; set; }
     public int? QuantityUnitId { get; set; }
 }
 
@@ -22,6 +24,6 @@ public class DeleteListIngredientsDto
 
 public record UpdateInventoryIngredientDto(
     string IngredientName,
-    int? Quantity,
+    decimal? Quantity,
     int? QuantityUnitId
 );
